flask_llm/requirements.txt
flask==3.1.2
gunicorn==22.0.0
langchain==0.3.27
langchain-openai==0.3.31
requests==2.32.5
python-dotenv==1.1.1
firecrawl-py
psycopg2-binary==2.9.10
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-Login==0.6.3
bcrypt==4.2.0
